--- app/app/src/main/java/dev/ksick/mapreactions/Place.java
@@ -1,13 +1,0 @@
-package dev.ksick.mapreactions;
-
-public class Place {
-    String name;
-    double latitude;
-    double longitude;
-
-    public Place(String name, double latitude, double longitude) {
-        this.name = name;
-        this.latitude = latitude;
-        this.longitude = longitude;
-    }
-}
