+++ app/app/src/main/java/dev/ksick/mapreactions/model/Place.java
@@ -1,0 +1,37 @@
+package dev.ksick.mapreactions.model;
+
+public class Place {
+    private String name;
+    private double latitude;
+    private double longitude;
+
+    public Place(String name, double latitude, double longitude) {
+        this.name = name;
+        this.latitude = latitude;
+        this.longitude = longitude;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    public double getLatitude() {
+        return latitude;
+    }
+
+    public void setLatitude(double latitude) {
+        this.latitude = latitude;
+    }
+
+    public double getLongitude() {
+        return longitude;
+    }
+
+    public void setLongitude(double longitude) {
+        this.longitude = longitude;
+    }
+}
