+++ app/app/src/main/res/layout/fragment_map.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MapFragment">
+    tools:context=".fragment.MapFragment">
 
     <org.osmdroid.views.MapView
         android:id="@+id/map_view"
