+++ app/app/src/main/res/layout/fragment_phrase_input.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@drawable/app_background"
-    tools:context=".PhraseInputFragment">
+    tools:context=".fragment.PhraseInputFragment">
 
     <View
         android:id="@+id/view"
