+++ app/app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/PhraseInputFragment"
-        android:name="dev.ksick.mapreactions.PhraseInputFragment"
+        android:name="dev.ksick.mapreactions.fragment.PhraseInputFragment"
         tools:layout="@layout/fragment_phrase_input">
 
         <action
@@ -16,7 +16,7 @@
     </fragment>
     <fragment
         android:id="@+id/MapFragment"
-        android:name="dev.ksick.mapreactions.MapFragment"
+        android:name="dev.ksick.mapreactions.fragment.MapFragment"
         tools:layout="@layout/fragment_map">
 
         <action
